+++ src/components/molecules/bookmarkCard/index.ts
@@ -1,0 +1,1 @@
+export * from './BookmarkCard';
